Refactor, extract flags classes out of superblock and inode class.
