+++ pom.xml
@@ -7,7 +7,7 @@
     <groupId>org.jnode</groupId>
     <artifactId>jnode-fs</artifactId>
     <packaging>jar</packaging>
-    <version>0.3.5-snapshot1</version>
+    <version>0.3.5-snapshot2</version>
 
     <dependencies>
         <dependency>
