+++ src/main/java/org/jnode/fs/xfs/XfsFileSystemType.java
@@ -1,14 +1,15 @@
 package org.jnode.fs.xfs;
 
+import java.io.IOException;
+import java.nio.ByteBuffer;
+
 import org.jnode.driver.Device;
 import org.jnode.driver.block.FSBlockDeviceAPI;
 import org.jnode.fs.BlockDeviceFileSystemType;
 import org.jnode.fs.FileSystemException;
+import org.jnode.fs.xfs.superblock.Superblock;
 import org.jnode.partitions.PartitionTableEntry;
 import org.jnode.util.BigEndian;
-
-import java.io.IOException;
-import java.nio.ByteBuffer;
 
 /**
  * The file system type for an XFS file system.
